Finished serialization using pugixml. Started to add the Chip/Component classes. Integrating them could get messy soon.

+++ cppsource/ViewResources.cpp
@@ -16,7 +16,7 @@ ViewResources::ViewResources()
 //Helper Functions
 sf::Vector2i mapCoordsToGrid(const sf::Vector2f & point)
 {
-    return sf::Vector2i{floor(point.x / GRID_SIZE.x), floor(point.y / GRID_SIZE.y)};
+    return sf::Vector2i{ static_cast<int>(floorf(point.x / GRID_SIZE.x)), static_cast<int>(floorf(point.y / GRID_SIZE.y)) };
 }
 
 sf::Vector2f mapGridToCoords(const sf::Vector2i & point)
